rename model files according to standards
