+++ apps/services/user-profile/src/app/user-profile/emailVerification.model.ts
