+++ apps/services/user-profile/src/app/user-profile/userProfile.module.ts
@@ -8,8 +8,8 @@ import { Logger, LOGGER_PROVIDER } from '@island.is/logging'
 import { DataSourceConfig } from 'apollo-datasource'
 import environment from '../../environments/environment'
 import { EmailService, EMAIL_OPTIONS } from '@island.is/email-service'
-import { SmsVerification } from './sms-verification.model'
-import { EmailVerification } from './email-verification.model'
+import { SmsVerification } from './smsVerification.model'
+import { EmailVerification } from './emailVerification.model'
 import { VerificationService } from './verification.service'
 import { UserProfileInfraController } from './userProfileInfra.controller'
 import { SequelizeConfigService } from '../sequelizeConfig.service'
@@ -48,4 +48,4 @@ import { SequelizeConfigService } from '../sequelizeConfig.service'
   ],
   exports: [UserProfileService],
 })
-export class UserProfileModule {}
+export class UserProfileModule { }
