+++ apps/services/user-profile/src/app/user-profile/verification.service.ts
@@ -1,12 +1,12 @@
 import { Logger, LOGGER_PROVIDER } from '@island.is/logging'
-import { BadRequestException, Inject, Injectable } from '@nestjs/common'
+import { Inject, Injectable } from '@nestjs/common'
 import { InjectModel } from '@nestjs/sequelize'
-import { EmailVerification } from './email-verification.model'
+import { EmailVerification } from './emailVerification.model'
 import * as CryptoJS from 'crypto-js'
 import { ConfirmEmailDto } from './dto/confirmEmailDto'
 import { UserProfile } from '../user-profile/userProfile.model'
 import { UserProfileService } from '../user-profile/userProfile.service'
-import { SmsVerification } from './sms-verification.model'
+import { SmsVerification } from './smsVerification.model'
 import { CreateUserProfileDto } from '../user-profile/dto/createUserProfileDto'
 import { SmsService } from '@island.is/nova-sms'
 import { EmailService } from '@island.is/email-service'
@@ -53,7 +53,7 @@ export class VerificationService {
     private readonly smsService: SmsService,
     @Inject(EmailService)
     private readonly emailService: EmailService,
-  ) {}
+  ) { }
 
   async createEmailVerification(
     nationalId: string,
